Refactor editContact function in ContactUseCase.
