+++ app/src/main/java/com/devmob/alaya/domain/ContactUseCase.kt
@@ -33,15 +33,34 @@ class ContactUseCase(private val prefs: SharedPreferences) {
         return contactRepositoryImpl.deleteContact(email, contact)
     }
 
-    suspend fun editContact(email: String, contact: Contact): FirebaseResult {
-        return contactRepositoryImpl.editContact(email, contact)
+    suspend fun editContact(contact: Contact): FirebaseResult {
+        val email = prefs.getEmail()
+        val user = email?.let { userRepository.getUser(it) }
+        val updatedContacts = user?.containmentNetwork?.toMutableList() ?: mutableListOf()
+        val index = updatedContacts.indexOfFirst { it.contactId == contact.contactId }
+
+        if (index != -1) {
+            val currentContact = updatedContacts[index]
+            var updatedContact = contact
+            if (!contact.photo.isNullOrEmpty() && contact.photo != currentContact.photo) {
+                val photoUri = Uri.parse(contact.photo)
+                val photoUrl = contactRepositoryImpl.uploadImageToStorage(photoUri, contact.contactId)
+                if (photoUrl != null) {
+                    updatedContact = contact.copy(photo = photoUrl)
+                } else {
+                    return FirebaseResult.Error(Exception("Error al subir la foto"))
+                }
+            }
+            updatedContacts[index] = updatedContact
+        }
+        return if (!email.isNullOrEmpty()) {
+            contactRepositoryImpl.updateContacts(email, updatedContacts)
+        } else {
+            FirebaseResult.Error(Throwable("Email is null"))
+        }
     }
 
     fun listenToContacts(email: String, onContactsUpdated: (List<Contact>) -> Unit) {
         contactRepositoryImpl.listenToContacts(email, onContactsUpdated)
-    }
-
-    suspend fun uploadImageToStorage(uri : Uri, contactId: String) {
-        contactRepositoryImpl.uploadImageToStorage(uri, contactId)
     }
 }
